Drop unused code from Mercurial
